fix(mock): gate the reversible server-fn id format to mock builds
`serverFns.generateFunctionId` was set unconditionally, so `bun run build` (the
one that produces the GHCR image) also emitted ids that are
base64url(JSON{file, export}) with no hash and no truncation. Every production
RPC URL therefore decoded back to a source path and an export name. Install the
option only when VITE_DEMO_MODE or VITE_ENABLE_MSW is set; production keeps
TanStack's opaque ids.

The encoder moves to src/lib/server-fn-id.ts so vite.config.ts no longer
statically imports a src/lib/mock module into the production build config.
src/lib/mock/handlers/function-id.ts re-exports it for the decoder side.
